fix(composer): rendering is now consistent between editor and deploy
