+++ assets/main.js
@@ -17,6 +17,9 @@ import { prefersReducedMotion, getArticleTitleFromMain } from './js/dom-utils.js
 import { renderPostMetaCard, renderOutdatedCard } from './js/templates.js';
 import { applyLangHints } from './js/typography.js';
 
+import { applyLazyLoadingIn, hydratePostImages, hydratePostVideos, hydrateCardCovers } from './js/post-render.js';
+import { hydrateInternalLinkCards } from './js/link-cards.js';
+
 // Lightweight fetch helper (bypass caches without version params)
 const getFile = (filename) => fetch(String(filename || ''), { cache: 'no-store' })
   .then(resp => { if (!resp.ok) throw new Error(`HTTP ${resp.status}`); return resp.text(); });
@@ -350,142 +353,7 @@ function renderSiteIdentity(cfg) {
   } catch (_) { /* noop */ }
 }
 
-// Ensure images defer offscreen loading for performance
-function applyLazyLoadingIn(container) {
-  try {
-    const root = typeof container === 'string' ? document.querySelector(container) : container;
-    if (!root) return;
-    const imgs = root.querySelectorAll('img');
-    imgs.forEach(img => {
-      if (!img.hasAttribute('loading')) img.setAttribute('loading', 'lazy');
-      if (!img.hasAttribute('decoding')) img.setAttribute('decoding', 'async');
-    });
-  } catch (_) {}
-}
-
 // Fade-in covers when each image loads; remove placeholder per-card
-function hydrateCardCovers(container) {
-  try {
-    const root = typeof container === 'string' ? document.querySelector(container) : (container || document);
-    if (!root) return;
-    const wraps = root.querySelectorAll('.index .card-cover-wrap, .link-card .card-cover-wrap');
-    wraps.forEach(wrap => {
-      const img = wrap.querySelector('img.card-cover');
-      if (!img) return;
-      const ph = wrap.querySelector('.ph-skeleton');
-      const done = () => {
-        img.classList.add('is-loaded');
-        if (ph && ph.parentNode) ph.parentNode.removeChild(ph);
-      };
-      if (img.complete && img.naturalWidth > 0) { done(); return; }
-      img.addEventListener('load', done, { once: true });
-      img.addEventListener('error', () => { if (ph && ph.parentNode) ph.parentNode.removeChild(ph); img.style.opacity = '1'; }, { once: true });
-      // Kick off loading only for link-card covers (index covers load sequentially elsewhere)
-      const inIndex = !!wrap.closest('.index');
-      if (!inIndex) {
-        // For link-card covers, src is set at creation time; no need to read from DOM attributes here.
-      } else {
-        // Index covers are handled by sequentialLoadCovers
-      }
-    });
-  } catch (_) {}
-}
-
-// Enhance post images: wrap with a reserved-ratio container + skeleton, fade-in when loaded
-function hydratePostImages(container) {
-  try {
-    const root = typeof container === 'string' ? document.querySelector(container) : (container || document);
-    if (!root) return;
-    const candidates = Array.from(root.querySelectorAll('img'))
-      .filter(img => !img.classList.contains('card-cover'))
-      .filter(img => !img.closest('table'))
-      .filter(img => !img.closest('figure'));
-  candidates.forEach(img => {
-      // Skip if already in a wrapper
-      if (img.closest('.post-image-wrap')) return;
-      // If the image lives inside a paragraph with other text, avoid restructuring
-      const p = img.parentElement && img.parentElement.tagName === 'P' ? img.parentElement : null;
-      const onlyThisImg = p ? (p.childElementCount === 1 && p.textContent.trim() === '') : false;
-
-      // Helper to create wrapper + skeleton and wire lazy fade-in
-      const setupWrap = (hostEl, nodeToMove) => {
-        const wrap = document.createElement('div');
-        wrap.className = 'post-image-wrap';
-        // Prefer explicit attributes for ratio if present
-        const wAttr = parseInt(img.getAttribute('width') || '', 10);
-        const hAttr = parseInt(img.getAttribute('height') || '', 10);
-        if (!isNaN(wAttr) && !isNaN(hAttr) && wAttr > 0 && hAttr > 0) {
-          wrap.style.aspectRatio = `${wAttr} / ${hAttr}`;
-        }
-        const ph = document.createElement('div');
-        ph.className = 'ph-skeleton';
-        ph.setAttribute('aria-hidden', 'true');
-
-        // Insert wrapper next to the image/link when in same parent; otherwise append
-        try {
-          if (nodeToMove && nodeToMove.parentElement === hostEl) hostEl.insertBefore(wrap, nodeToMove);
-          else hostEl.appendChild(wrap);
-        } catch (_) { hostEl.appendChild(wrap); }
-        wrap.appendChild(ph);
-        wrap.appendChild(nodeToMove || img);
-
-        img.classList.add('post-img');
-        if (!img.hasAttribute('loading')) img.setAttribute('loading', 'lazy');
-        if (!img.hasAttribute('decoding')) img.setAttribute('decoding', 'async');
-
-        const src = img.getAttribute('src');
-        if (src) { img.setAttribute('data-src', src); img.removeAttribute('src'); }
-
-        const done = () => {
-          if (img.naturalWidth && img.naturalHeight) {
-            wrap.style.aspectRatio = `${img.naturalWidth} / ${img.naturalHeight}`;
-          }
-          img.classList.add('is-loaded');
-          if (ph && ph.parentNode) ph.parentNode.removeChild(ph);
-        };
-        if (img.complete && img.naturalWidth > 0) { done(); }
-        else {
-          img.addEventListener('load', done, { once: true });
-          img.addEventListener('error', () => { if (ph && ph.parentNode) ph.parentNode.removeChild(ph); img.style.opacity = '1'; }, { once: true });
-        }
-
-        if (src) {
-          const safe = sanitizeImageUrl(src);
-          if (safe) img.src = safe;
-        }
-      };
-
-      if (onlyThisImg) {
-        // Upgrade to a semantic figure with figcaption using alt text
-        const alt = (img.getAttribute('alt') || '').trim();
-        const figure = document.createElement('figure');
-        // Insert figure before the paragraph, then move image inside
-        p.parentElement && p.parentElement.insertBefore(figure, p);
-        // If image is wrapped in an anchor directly under <p>, move the anchor as a unit
-        const link = (img.parentElement && img.parentElement.tagName === 'A' && img.parentElement.parentElement === p) ? img.parentElement : null;
-        const nodeToMove = link || img;
-        // Put skeleton wrapper inside figure and move node (img or link)
-        setupWrap(figure, nodeToMove);
-        // Remove empty paragraph shell
-        try { if (p.parentElement) p.parentElement.removeChild(p); } catch (_) {}
-        // Add caption only if we have non-empty alt
-        if (alt) {
-          const cap = document.createElement('figcaption');
-          cap.textContent = alt;
-          figure.appendChild(cap);
-        }
-      } else {
-        // Regular inline/embedded image: wrap in a reserved-ratio container only
-        const targetParent = img.parentElement;
-        if (!targetParent) return;
-        // Move any wrapping anchor along with the image if present
-        const link = (img.parentElement && img.parentElement.tagName === 'A') ? img.parentElement : null;
-        const nodeToMove = link || img;
-        setupWrap(targetParent, nodeToMove);
-      }
-    });
-  } catch (_) {}
-}
 // --- Asset watchdog: warn when image assets exceed configured threshold ---
 async function checkImageSize(url, timeoutMs = 4000) {
   // Try HEAD first; fall back to range request when HEAD not allowed
@@ -571,421 +439,7 @@ async function warnLargeImagesIn(container, cfg = {}) {
 }
 
 
-// Auto-generate preview posters for videos to avoid gray screen before play
-function hydratePostVideos(container) {
-  try {
-    const root = typeof container === 'string' ? document.querySelector(container) : (container || document);
-    if (!root) return;
-    // Normalize raw HTML <video> markup to match NanoSite's generated structure
-    const videos = Array.from(root.querySelectorAll('video'));
-    videos.forEach(video => {
-      try {
-        if (!video.classList.contains('post-video')) video.classList.add('post-video');
-        if (!video.hasAttribute('controls')) video.setAttribute('controls', '');
-        if (!video.hasAttribute('preload')) video.setAttribute('preload', 'metadata');
-        if (!video.hasAttribute('playsinline')) video.setAttribute('playsinline', '');
-        if (!video.closest('.post-video-wrap')) {
-          const wrap = document.createElement('div');
-          wrap.className = 'post-video-wrap';
-          const parent = video.parentElement;
-          if (parent) {
-            parent.insertBefore(wrap, video);
-            wrap.appendChild(video);
-          }
-        }
-        // Ensure browser re-evaluates sources after moving/normalizing
-        try { video.load(); } catch (_) {}
-
-        // Install a lightweight fallback overlay on error
-        try {
-          if (video.dataset.vfInstalled !== '1') {
-            const wrap = video.closest('.post-video-wrap') || video.parentElement;
-            const overlay = document.createElement('div');
-            overlay.className = 'video-fallback';
-            overlay.style.display = 'none';
-            const toAbs = (s) => { try { return new URL(s, window.location.href).toString(); } catch(_) { return s; } };
-            const sources = [video.getAttribute('src'), ...Array.from(video.querySelectorAll('source')).map(s => s.getAttribute('src'))].filter(Boolean);
-            const first = sources.length ? toAbs(sources[0]) : '#';
-            // Build fallback overlay using DOM APIs to avoid HTML injection
-            const content = document.createElement('div');
-            content.className = 'vf-content';
-            const title = document.createElement('div');
-            title.className = 'vf-title';
-            title.textContent = 'Video not available';
-            const actions = document.createElement('div');
-            actions.className = 'vf-actions';
-            const link = document.createElement('a');
-            link.className = 'vf-link primary';
-            try { link.setAttribute('rel', 'noopener'); } catch(_) {}
-            link.setAttribute('target', '_blank');
-            const sanitized = sanitizeUrl(first);
-            link.setAttribute('href', sanitized || '#');
-            link.textContent = 'Open file';
-            actions.appendChild(link);
-            content.appendChild(title);
-            content.appendChild(actions);
-            overlay.appendChild(content);
-            if (wrap && !wrap.querySelector('.video-fallback')) wrap.appendChild(overlay);
-            const show = () => { overlay.style.display = 'flex'; };
-            const hide = () => { overlay.style.display = 'none'; };
-            video.addEventListener('error', show);
-            video.addEventListener('loadeddata', hide);
-            video.addEventListener('canplay', hide);
-            video.addEventListener('play', hide);
-            // If all sources error, show overlay soon
-            Array.from(video.querySelectorAll('source')).forEach(s => { s.addEventListener('error', show); });
-            video.dataset.vfInstalled = '1';
-          }
-        } catch(_) {}
-      } catch (_) {}
-    });
-    const queue = Array.from(root.querySelectorAll('video'))
-      .filter(video => video && !video.hasAttribute('poster') && video.dataset.autoposterDone !== '1');
-
-    const delay = (ms) => new Promise(r => setTimeout(r, ms));
-
-    const waitForMetadata = (video) => new Promise(resolve => {
-      if (!video) return resolve();
-      if (video.readyState >= 2) return resolve(); // HAVE_CURRENT_DATA is safest for draw
-      const onMeta = () => { /* keep waiting for data */ };
-      const onData = () => { cleanup(); resolve(); };
-      const onErr = () => { cleanup(); resolve(); };
-      const cleanup = () => {
-        video.removeEventListener('loadedmetadata', onMeta);
-        video.removeEventListener('loadeddata', onData);
-        video.removeEventListener('error', onErr);
-      };
-      video.addEventListener('loadedmetadata', onMeta, { once: true });
-      video.addEventListener('loadeddata', onData, { once: true });
-      video.addEventListener('error', onErr, { once: true });
-      // Safety timeout: if we only get metadata, proceed after 1200ms
-      setTimeout(() => { cleanup(); resolve(); }, 1200);
-    });
-
-    const LITE_POSTER = true; // Avoid seeks/play to improve stability
-    const processVideo = async (video) => {
-      try {
-        // Skip if already set
-        if (!video || video.hasAttribute('poster') || video.dataset.autoposterDone === '1') return;
-        // Ensure minimal attributes for better UX
-  if (!video.hasAttribute('preload')) video.setAttribute('preload', 'metadata');
-  if (!video.hasAttribute('playsinline')) video.setAttribute('playsinline', '');
-
-        await waitForMetadata(video);
-
-  // Snapshot original state to safely restore after probing
-  const origTime = (() => { try { return Number(video.currentTime) || 0; } catch(_) { return 0; } })();
-  const wasPaused = !!video.paused;
-
-        const drawPoster = () => {
-          try {
-            const w = Math.max(1, Number(video.videoWidth) || 0);
-            const h = Math.max(1, Number(video.videoHeight) || 0);
-            if (!w || !h) return false;
-            const canvas = document.createElement('canvas');
-            canvas.width = w; canvas.height = h;
-            const ctx = canvas.getContext('2d');
-            if (!ctx) return false;
-            ctx.drawImage(video, 0, 0, w, h);
-            // Reject too-dark frames
-            try {
-              const small = document.createElement('canvas');
-              small.width = 16; small.height = 16;
-              const sctx = small.getContext('2d');
-              if (sctx) {
-                sctx.drawImage(canvas, 0, 0, 16, 16);
-                const img = sctx.getImageData(0, 0, 16, 16);
-                let sum = 0, n = img.data.length / 4;
-                for (let i = 0; i < img.data.length; i += 4) {
-                  const r = img.data[i], g = img.data[i+1], b = img.data[i+2];
-                  sum += 0.2126*r + 0.7152*g + 0.0722*b;
-                }
-                const avg = sum / n;
-                if (avg < 10) return false;
-              }
-            } catch(_) {}
-            const dataUrl = canvas.toDataURL('image/jpeg', 0.84);
-            if (dataUrl && dataUrl.startsWith('data:image')) {
-              video.setAttribute('poster', dataUrl);
-              video.dataset.autoposterDone = '1';
-              return true;
-            }
-          } catch (_) {}
-          return false;
-        };
-
-        const captureAt = (t) => new Promise((resolve) => {
-          const cleanup = () => {
-            video.removeEventListener('seeked', onSeeked);
-            video.removeEventListener('error', onError);
-          };
-          const onSeeked = () => {
-            const ok = drawPoster();
-            cleanup();
-            // Restore position to avoid leaving the media in a probed time
-            try { if (Number.isFinite(origTime)) video.currentTime = origTime; } catch(_) {}
-            resolve(ok);
-          };
-          const onError = () => {
-            cleanup();
-            try { if (Number.isFinite(origTime)) video.currentTime = origTime; } catch(_) {}
-            resolve(false);
-          };
-          video.addEventListener('seeked', onSeeked, { once: true });
-          video.addEventListener('error', onError, { once: true });
-          try {
-            // Only seek to times within seekable ranges
-            const seekable = video.seekable;
-            if (!seekable || seekable.length === 0) { cleanup(); resolve(false); return; }
-            let target = Number.isFinite(t) ? t : 0;
-            let clamped = false;
-            for (let i = 0; i < seekable.length; i++) {
-              const start = seekable.start(i);
-              const end = seekable.end(i);
-              if (target >= start && target <= end) { clamped = true; break; }
-            }
-            if (!clamped) {
-              const start = seekable.start(0);
-              // Push a small epsilon inside the range
-              target = Math.min(seekable.end(0) - 0.01, start + 0.12);
-            }
-            if (!Number.isFinite(target) || target < 0) { cleanup(); resolve(false); return; }
-            video.currentTime = target;
-          } catch (_) {
-            cleanup(); resolve(false);
-          }
-        });
-
-        const captureWithRVFC = () => new Promise((resolve) => {
-          try {
-            if (!('requestVideoFrameCallback' in HTMLVideoElement.prototype)) {
-              resolve(false); return;
-            }
-            const rvfc = video.requestVideoFrameCallback.bind(video);
-            const id = rvfc(() => { const ok = drawPoster(); resolve(!!ok); });
-            setTimeout(() => { try { video.cancelVideoFrameCallback && video.cancelVideoFrameCallback(id); } catch(_){} resolve(false); }, 450);
-          } catch (_) { resolve(false); }
-        });
-
-    const captureByPlayPause = async () => {
-          const wasPaused = video.paused;
-          const wasMuted = video.muted;
-          try {
-      // Only attempt if user has interacted, to avoid blocked autoplay states
-      if (navigator.userActivation && !navigator.userActivation.hasBeenActive) return false;
-            video.muted = true;
-            await (video.play && video.play().catch(() => {}));
-            await delay(180);
-            const ok = drawPoster();
-            if (ok) return true;
-          } catch (_) {}
-          finally {
-            try { video.pause && video.pause(); } catch(_) {}
-      try { if (Number.isFinite(origTime)) video.currentTime = origTime; } catch(_) {}
-            video.muted = wasMuted;
-            if (!wasPaused) { try { await video.play(); } catch(_) {} }
-          }
-          return false;
-        };
-
-        // Attempt strategies
-        // Build offsets inside first seekable range
-        const offsets = [];
-        const seekable = video.seekable;
-        const hasRange = seekable && seekable.length > 0;
-        const rStart = hasRange ? seekable.start(0) : 0;
-        const rEnd = hasRange ? seekable.end(0) : Math.max(1, Number(video.duration) || 1);
-        const dur = Math.max(0, rEnd - rStart);
-        const base = dur > 0.25 ? Math.min(rStart + 0.25, rStart + dur / 12) : rStart + 0.14;
-        offsets.push(base);
-        if (dur) {
-          offsets.push(Math.min(rEnd - 0.5, rStart + dur * 0.05));
-          offsets.push(Math.min(rEnd - 0.1, rStart + dur * 0.1));
-        }
-
-        if (drawPoster()) return;
-        if (await captureWithRVFC()) return;
-        if (LITE_POSTER) return; // Skip aggressive strategies in lite mode
-        // Avoid aggressive seeks on iOS Safari or QuickTime MOV sources
-        const ua = navigator.userAgent || '';
-        const isIOS = /iP(hone|ad|od)/.test(ua) || (/Mac/.test(ua) && 'ontouchend' in document);
-        const srcUrl = String(video.currentSrc || (video.querySelector('source') && video.querySelector('source').src) || '').toLowerCase();
-        const isMov = srcUrl.endsWith('.mov') || srcUrl.includes('video/quicktime');
-        if (isMov) {
-          // Skip auto poster for MOV/QuickTime to avoid WebKit decode issues
-        } else if (!isIOS) {
-          for (const off of offsets) {
-            const ok = await captureAt(off);
-            if (ok) { return; }
-          }
-          const ok0 = await captureAt(0);
-          if (ok0) { return; }
-        }
-        await captureByPlayPause();
-
-        // Ensure final state is sane
-        try { if (Number.isFinite(origTime)) video.currentTime = origTime; } catch(_) {}
-        if (!wasPaused) { try { await video.play(); } catch(_) {} }
-      } catch (_) { /* ignore per-video errors */ }
-    };
-
-    (async () => {
-      for (const v of queue) {
-        await processVideo(v);
-        // brief gap between videos to reduce decoder contention
-        await delay(80);
-      }
-    })();
-  } catch (_) {}
-}
-
-
 // Transform standalone internal links (?id=...) into rich article cards
-function hydrateInternalLinkCards(container) {
-  try {
-    const root = typeof container === 'string' ? document.querySelector(container) : (container || document);
-    if (!root) return;
-    const anchors = Array.from(root.querySelectorAll('a[href^="?id="]'));
-    if (!anchors.length) return;
-
-    const isWhitespaceOnlySiblings = (el) => {
-      const p = el && el.parentNode;
-      if (!p) return false;
-      const nodes = Array.from(p.childNodes || []);
-      return nodes.every(n => (n === el) || (n.nodeType === Node.TEXT_NODE && !String(n.textContent || '').trim()));
-    };
-
-    const parseId = (href) => {
-      try { const u = new URL(href, window.location.href); return u.searchParams.get('id'); } catch (_) { return null; }
-    };
-
-    // Simple cache to avoid refetching the same markdown multiple times per page
-    const mdCache = new Map(); // location -> Promise<string>
-
-    anchors.forEach(a => {
-      const rawLoc = parseId(a.getAttribute('href') || '');
-      if (!rawLoc) return;
-      // Prefer current-language alias if available (e.g., link points to main_en.md but UI is zh)
-      const aliased = (locationAliasMap && locationAliasMap.has(rawLoc)) ? locationAliasMap.get(rawLoc) : rawLoc;
-      // Allow either the raw location or its alias (covers cross-language links)
-      if (!allowedLocations.has(rawLoc) && !allowedLocations.has(aliased)) return;
-      const loc = aliased;
-
-      // Only convert when link is the only content in its block container (p/li/div)
-      const parent = a.parentElement;
-      const isStandalone = parent && ['P', 'LI', 'DIV'].includes(parent.tagName) && isWhitespaceOnlySiblings(a);
-      const titleAttr = (a.getAttribute('title') || '').trim();
-      const forceCard = /\b(card|preview)\b/i.test(titleAttr) || a.hasAttribute('data-card') || a.classList.contains('card');
-      if (!isStandalone && !forceCard) return;
-
-      // Lookup metadata from loaded index cache
-  const title = postsByLocationTitle[loc] || loc;
-  const meta = (Object.entries(postsIndexCache || {}) || []).find(([, v]) => v && v.location === loc)?.[1] || {};
-  const href = withLangParam(`?id=${encodeURIComponent(loc)}`);
-      const tagsHtml = meta ? renderTags(meta.tag) : '';
-      const dateHtml = meta && meta.date ? `<span class="card-date">${escapeHtml(formatDisplayDate(meta.date))}</span>` : '';
-      const draftHtml = meta && meta.draft ? `<span class="card-draft">${t('ui.draftBadge')}</span>` : '';
-      // Allow relative frontmatter image (e.g., 'cover.jpg'); resolve against the post's folder
-      const rawCover = meta && (meta.thumb || meta.cover || meta.image);
-      let coverSrc = rawCover;
-  if (rawCover && typeof rawCover === 'string' && !/^https?:\/\//i.test(rawCover) && !rawCover.startsWith('/') && !rawCover.includes('/')) {
-        const baseLoc = (meta && meta.location) || loc; // use current link's location as base
-        const lastSlash = String(baseLoc || '').lastIndexOf('/');
-        const baseDir = lastSlash >= 0 ? String(baseLoc).slice(0, lastSlash + 1) : '';
-        coverSrc = (baseDir + rawCover).replace(/\/+/, '/');
-      }
-      const useFallbackCover = !(siteConfig && siteConfig.cardCoverFallback === false);
-      const cover = (coverSrc)
-        ? `<div class="card-cover-wrap"><div class="ph-skeleton" aria-hidden="true"></div><img class="card-cover" alt="${escapeHtml(title)}" src="${escapeHtml(cardImageSrc(coverSrc))}" loading="lazy" decoding="async" fetchpriority="low" width="1600" height="1000"></div>`
-        : (useFallbackCover ? fallbackCover(title) : '');
-
-      const wrapper = document.createElement('div');
-      wrapper.className = 'link-card-wrap';
-      const initialMeta = [dateHtml, draftHtml].filter(Boolean).join('<span class="card-sep">•</span>');
-      wrapper.innerHTML = `<a class="link-card" href="${href}">${cover}<div class="card-title">${escapeHtml(title)}</div><div class="card-excerpt">${t('ui.loading')}</div><div class="card-meta">${initialMeta}</div>${tagsHtml}</a>`;
-
-      // If index metadata provides an explicit excerpt, prefer it immediately
-      try {
-        const exNode = wrapper.querySelector('.card-excerpt');
-        if (exNode && meta && meta.excerpt) {
-          exNode.textContent = String(meta.excerpt);
-        }
-      } catch (_) {}
-
-      // Placement rules:
-      // - If standalone in LI: replace the anchor to keep list structure
-      // - If standalone in P/DIV: replace the container with the card
-      // - If forced (title contains 'card' or similar) but not standalone:
-      //   insert the card right after the parent block, remove the anchor;
-      //   if the parent becomes empty, remove it too.
-      if (parent.tagName === 'LI' && isStandalone) {
-        a.replaceWith(wrapper);
-      } else if (isStandalone && (parent.tagName === 'P' || parent.tagName === 'DIV')) {
-        const target = parent;
-        target.parentNode.insertBefore(wrapper, target);
-        target.remove();
-      } else {
-        // forced-card, inline inside a block
-        const after = parent.nextSibling;
-        parent.parentNode.insertBefore(wrapper, after);
-        // remove the anchor from inline text
-        a.remove();
-        // if paragraph becomes empty/whitespace, remove it
-        if (!parent.textContent || !parent.textContent.trim()) {
-          parent.remove();
-        }
-      }
-
-      // Lazy-hydrate cover image
-      hydrateCardCovers(wrapper);
-
-      // Fetch markdown to compute excerpt + read time
-      const ensureMd = (l) => {
-        if (!mdCache.has(l)) mdCache.set(l, getFile(`${getContentRoot()}/${l}`).catch(() => ''));
-        return mdCache.get(l);
-      };
-      ensureMd(loc).then(md => {
-        if (!wrapper.isConnected) return;
-        const ex = extractExcerpt(md, 50);
-        const minutes = computeReadTime(md, 200);
-        const card = wrapper.querySelector('a.link-card');
-        if (!card) return;
-        const exEl = card.querySelector('.card-excerpt');
-  // Only override excerpt if no explicit excerpt in metadata
-  if (exEl && !(meta && meta.excerpt)) exEl.textContent = ex;
-        const metaEl = card.querySelector('.card-meta');
-        if (metaEl) {
-          // Rebuild meta using DOM nodes instead of HTML strings
-          const items = [];
-          const dateEl = metaEl.querySelector('.card-date');
-          if (dateEl && dateEl.textContent.trim()) items.push(dateEl.cloneNode(true));
-          const read = document.createElement('span');
-          read.className = 'card-read';
-          read.textContent = `${minutes} ${t('ui.minRead')}`;
-          items.push(read);
-          if (meta && meta.draft) {
-            const d = document.createElement('span');
-            d.className = 'card-draft';
-            d.textContent = t('ui.draftBadge');
-            items.push(d);
-          }
-          // Clear and append with separators
-          metaEl.textContent = '';
-          items.forEach((node, idx) => {
-            if (idx > 0) {
-              const sep = document.createElement('span');
-              sep.className = 'card-sep';
-              sep.textContent = '•';
-              metaEl.appendChild(sep);
-            }
-            metaEl.appendChild(node);
-          });
-        }
-      }).catch(() => {});
-    });
-  } catch (_) {}
-}
-
 // Load cover images sequentially to reduce bandwidth contention
 function sequentialLoadCovers(container, maxConcurrent = 1) {
   try {
@@ -1539,7 +993,16 @@ function displayPost(postname) {
       const cfg = (siteConfig && siteConfig.assetWarnings && siteConfig.assetWarnings.largeImage) || {};
       warnLargeImagesIn('#mainview', cfg);
     } catch (_) {}
-  try { hydrateInternalLinkCards('#mainview'); } catch (_) {}
+  try { hydrateInternalLinkCards('#mainview', {
+    allowedLocations,
+    locationAliasMap,
+    postsByLocationTitle,
+    postsIndexCache,
+    siteConfig,
+    translate: t,
+    makeHref: (loc) => withLangParam(`?id=${encodeURIComponent(loc)}`),
+    fetchMarkdown: (loc) => getFile(`${getContentRoot()}/${loc}`)
+  }); } catch (_) {}
   try { hydratePostVideos('#mainview'); } catch (_) {}
   // Wire up copy-link buttons on all post meta cards
   try {
@@ -2012,7 +1475,16 @@ function displayStaticTab(slug) {
         const cfg = (siteConfig && siteConfig.assetWarnings && siteConfig.assetWarnings.largeImage) || {};
         warnLargeImagesIn('#mainview', cfg);
       } catch (_) {}
-  try { hydrateInternalLinkCards('#mainview'); } catch (_) {}
+  try { hydrateInternalLinkCards('#mainview', {
+    allowedLocations,
+    locationAliasMap,
+    postsByLocationTitle,
+    postsIndexCache,
+    siteConfig,
+    translate: t,
+    makeHref: (loc) => withLangParam(`?id=${encodeURIComponent(loc)}`),
+    fetchMarkdown: (loc) => getFile(`${getContentRoot()}/${loc}`)
+  }); } catch (_) {}
   try { hydratePostVideos('#mainview'); } catch (_) {}
   try { initSyntaxHighlighting(); } catch (_) {}
   try { renderTagSidebar(postsIndexCache); } catch (_) {}
